requirements.txt
fastapi
uvicorn
gitpython
networkx
tree-sitter
tree-sitter-languages
python-multipart
python-dotenv
faiss-cpu
numpy
sentence-transformers

ollama